initial steps to the EVM-CEP integration
The integration test case captures the usage of the CEP patterns,
CEP-RuleSpecifications and CEP-RuleInstances from the aspect of the
user, i.e. the sources to be generated are collected into the src-gen
folder, the one written by the user are in the src folder.
In order to facilitate registering CepRuleSpecifications, some
modifications have been made on the EventModelManager class.